decrypt_file
